+++ CHANGELOG.md
@@ -6,7 +6,7 @@ General
 
 - Several `let` declarations are no mutable
 
-Helpers
+Architectural Pattern Helpers
 
 - `UIAlertViewModel` has a new API that supports many buttons with state and style