+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Helpers
 
 Architectural Pattern Helpers
 
+- `UIActivityIndicatorViewable` now has a default implementation for `UIView`
 - `UIAlertViewModel` has a new API that supports many buttons with state and style
 
 #### [3.3.1(24)](https://github.com/VakhoKontridze/VCore/releases/tag/3.3.1) — *2022 06 08*