+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Architectural Pattern Helpers
 - `UIActivityIndicatorViewable` now has a default implementation for `UIView`
 - `UIAlertViewModel` has a new API that supports many buttons with state and style
 
+Extensions
+
+- `Optional` `String` method `compactMapNonEmpty` is renamed to `compactMapNonNilNonEmpty`
+
 #### [3.3.1(24)](https://github.com/VakhoKontridze/VCore/releases/tag/3.3.1) — *2022 06 08*
 
 Helpers