+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-#### [3.4.0(25)]
+#### [3.4.0(25)](https://github.com/VakhoKontridze/VCore/releases/tag/3.4.0) — *2022 06 11*
 
 General